+++ soffa-foundation-core/src/main/java/io/soffa/foundation/core/actions/MessageHandler.java
@@ -12,4 +12,11 @@ public interface MessageHandler {
      * @param message The message to handle
      */
     Optional<Object> onMessage(Message message);
+
+    default boolean accept(String action) {
+        return true;
+    }
+
+
+
 }