+++ soffa-foundation-test-app/src/test/java/com/company/app/events/EventHandleTest.java
@@ -1,7 +1,7 @@
 package com.company.app.events;
 
 import io.soffa.foundation.context.TenantHolder;
-import io.soffa.foundation.core.actions.MessageHandler;
+import io.soffa.foundation.core.actions.MessagesHandler;
 import io.soffa.foundation.core.data.SysLogRepository;
 import io.soffa.foundation.core.messages.Message;
 import io.soffa.foundation.exceptions.FakeException;
@@ -22,7 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class EventHandleTest {
 
     @Autowired
-    private MessageHandler messageHandler;
+    private MessagesHandler handler;
 
     @Autowired
     private SysLogRepository sysLogs;
@@ -34,9 +34,9 @@ public class EventHandleTest {
 
         TenantHolder.use("T1", (t1) -> {
             long initialCount = sysLogs.count();
-            messageHandler.onMessage(new Message(actionName)); // automatic tenant
-            messageHandler.onMessage(new Message("EchoAction", "Hello"));
-            messageHandler.onMessage(new Message(actionName).withTenant(t1)); // explicit tenant
+            handler.onMessage(new Message(actionName)); // automatic tenant
+            handler.onMessage(new Message("EchoAction", "Hello"));
+            handler.onMessage(new Message(actionName).withTenant(t1)); // explicit tenant
             Awaitility.await().atMost(500, TimeUnit.MILLISECONDS).until(() -> sysLogs.count() == initialCount + 3);
         });
 
@@ -44,7 +44,7 @@ public class EventHandleTest {
 
         TenantHolder.use("T2", (t2) -> {
             t2InitialCount.set(sysLogs.count());
-            Assertions.assertThrows(FakeException.class, () -> messageHandler.onMessage(new Message(actionName)));
+            Assertions.assertThrows(FakeException.class, () -> handler.onMessage(new Message(actionName)));
         });
 
         TenantHolder.use("T2", () -> {